+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Refreshed versions and urls.
   [sgeulette]
+- Added AssignedUsersVocabulary for assigned_user field so this vocabulary can be overrided.
+  [sgeulette]
 
 3.0.9 (2022-06-21)
 ------------------